Modified commands
Replaced !facts with !fact
Removed !sam
