+++ commands/misc.py
@@ -7,8 +7,7 @@ GO_HELP_TEXT = """The eternal #cs meme."""
 DUNNO_HELP_TEXT = """¯\\_(ツ)_/¯"""
 RUST_HELP_TEXT = """And if you gaze long into RUST, the RUST also gazes into you."""
 PR_HELP_TEXT = """You know what to do"""
-BLUESHELL_HELP_TEXT = """Thank you RNGesus for the £5 donation!"""
-SAM_HELP_TEXT = """Sam."""
+BLUESHELL_HELP_TEXT = """!blueshell"""
 
 
 class Misc:
@@ -41,11 +40,7 @@ class Misc:
 
     @commands.command(help=BLUESHELL_HELP_TEXT, brief=BLUESHELL_HELP_TEXT)
     async def blueshell(self, ctx: Context):
-        await ctx.send(":blueshell:")
-
-    @commands.command(help=SAM_HELP_TEXT, brief=SAM_HELP_TEXT)
-    async def sam(self, ctx: Context):
-        await ctx.send(":playingCoy:")
+        await ctx.send("Thank you RNGesus for the £5 donation!")
 
 
 def setup(bot: Bot):
